Committing files,
about to start adding in new functionality to profile view, they include.

Making a list of notifications the user can set so that they can be reminded before or after training to log progress.

Allowing the user to control when they get notifications from the app and not the other way around.

I was thinking... I could do this via location, but thats just another thing I have to track from my users.