Did a full redesign of the stat sheet. I am getting the app ready for TestFlight mode. currently it doesn't check if the users email is really their email.
I am going to add that in after i know it works functionally.
I have also added in some app icons.
added the ability for the user to delete their account. But.. there is no warning for the user if they are to delete it yet.
Currently the app seems like there is nothing wrong just wanting to get people to test and see what needs to be done.