A few small changes have been made.
I have added in a bunch more submissions for people to log. It has grown from only three of each submission type.
There are also a few little improvements I have tried to add, I tried to improve the navigation of the new submissions when you switch through the different submission types(Still working on it)
Need to improve on the loading submission once someone logs in.
Need to improve the loading and viewing of the stats view. It looks kinda jumpy.
Also added a few small things to the reminder view.

Other than that the beta seems to be going well.