Currently working on the simulator. Still needs some more testing on an iPhone
I have the user logging out and then terminating the app, re-logging back in and then loading up all their submissions they have previously recored.

Before it would log them in they would then need to terminate the app to be able to get their submission to show up.

I have also implemented a pull down function to reload the list and pull in all of the submissions the user has recored.

I think next would be to show the user if they have submission to fetch or if they haven't recorded anything.

Just some small things i am adding in to get MyBJJ free version ready to be launched onto the app store. The Submission library is still a work in progress. I think i will still keep it in for the app store launch, but letting users know more is to come. That is the same with the navigation link when someone is to tap on a submission in the list of recored submissions. It will currently show them nothing, or hey I am working on this hang tight please :)